Corrijo update de platos

+++ app/controllers/api/v1/Plates.java
@@ -100,13 +100,15 @@ public class Plates extends Controller {
 
             if(formData != null){
                 Http.MultipartFormData.FilePart pictureFilePart = formData.getFile("picture");
-                if(FolderServices.fileExists(FolderServices.getCommerceFolder(commerceUser.getCommerce()) + pictureFilePart.getFilename()) && !Plate.findByProperty("id", id).getPictureFileName().equals(pictureFilePart.getFilename())){
-                    logger.error("Imagen con nombre ya utilizado");
-                    return badRequest(JsonNodeFactory.instance.objectNode().put("message", "Nombre de imagen ya utilizado"));
+                if(pictureFilePart != null) {
+                    if (FolderServices.fileExists(FolderServices.getCommerceFolder(commerceUser.getCommerce()) + pictureFilePart.getFilename()) && !Plate.findByProperty("id", id).getPictureFileName().equals(pictureFilePart.getFilename())) {
+                        logger.error("Imagen con nombre ya utilizado");
+                        return badRequest(JsonNodeFactory.instance.objectNode().put("message", "Nombre de imagen ya utilizado"));
+                    }
+                    //Elimino el archivo imagen y creo el nuevo
+                    new File(FolderServices.getCommerceFolder(commerceUser.getCommerce()) + Plate.findByProperty("id", id).getPictureFileName()).delete();
+                    pictureFilePart.getFile().renameTo(new File(FolderServices.getCommerceFolder(commerceUser.getCommerce()) + pictureFilePart.getFilename()));
                 }
-                //Elimino el archivo imagen y creo el nuevo
-                new File(FolderServices.getCommerceFolder(commerceUser.getCommerce()) + Plate.findByProperty("id", id).getPictureFileName()).delete();
-                pictureFilePart.getFile().renameTo(new File(FolderServices.getCommerceFolder(commerceUser.getCommerce()) + pictureFilePart.getFilename()));
             }
             //Se actualiza el plato para el comercio especificado
             PlatesServices.update(id, plate, commerceUser.getCommerce());
